+++ noodles-cram/examples/cram_view.rs
@@ -19,14 +19,17 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let src = args.next().expect("missing src");
     let fasta_src = args.next();
 
-    let repository = fasta_src
+    let reference_sequence_repository = fasta_src
         .map(|src| fasta::indexed_reader::Builder::default().build_from_path(src))
         .transpose()?
         .map(IndexedReader::new)
         .map(fasta::Repository::new)
         .unwrap_or_default();
 
-    let mut reader = cram::reader::Builder::default().build_from_path(src)?;
+    let mut reader = cram::reader::Builder::default()
+        .set_reference_sequence_repository(reference_sequence_repository)
+        .build_from_path(src)?;
+
     reader.read_file_definition()?;
 
     let header = reader.read_file_header()?;
@@ -34,7 +37,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let stdout = io::stdout().lock();
     let mut writer = sam::Writer::new(BufWriter::new(stdout));
 
-    for result in reader.records(&repository, &header) {
+    for result in reader.records(&header) {
         let record = result.and_then(|record| record.try_into_alignment_record(&header))?;
         writer.write_alignment_record(&header, &record)?;
     }